TST: Fix squelcher test

+++ intelmq/tests/bots/experts/squelcher/test_expert.py
@@ -100,20 +100,19 @@ class TestSquelcherExpertBot(test.BotTestCase, unittest.TestCase):
     def set_bot(cls):
         cls.bot_reference = SquelcherExpertBot
         cls.default_input_message = INPUT1
-        try:
-            cls.sysconfig = (utils.load_configuration(RUNTIME_CONF_FILE)
-                             ['Expert']['Squelcher'])
-        except:
-            cls.sysconfig = {"configuration_path": pkg_resources.resource_filename('intelmq',
-                                                                                   'etc/squelcher.conf'),
-                             "host": "localhost",
-                             "port": 5432,
-                             "database": "intelmq",
-                             "user": "intelmq",
-                             "password": "intelmq",
-                             "sslmode": "require",
-                             "table": "tests",
-                             }
+        if not os.environ.get('INTELMQ_TEST_DATABASES'):
+            return
+        cls.sysconfig = {"configuration_path": pkg_resources.resource_filename('intelmq',
+                                                                               'etc/squelcher.conf'),
+                         "host": "localhost",
+                         "port": 5432,
+                         "database": "intelmq",
+                         "user": "intelmq",
+                         "password": "intelmq",
+                         "sslmode": "require",
+                         "table": "tests",
+                         "logging_level": "DEBUG",
+                         }
         cls.con = psycopg2.connect(database=cls.sysconfig['database'],
                                    user=cls.sysconfig['user'],
                                    password=cls.sysconfig['password'],
@@ -230,6 +229,8 @@ class TestSquelcherExpertBot(test.BotTestCase, unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
+        if not os.environ.get('INTELMQ_TEST_DATABASES'):
+            return
         cls.cur.execute("TRUNCATE TABLE {}".format(cls.sysconfig['table']))
         cls.cur.close()
         cls.con.close()
